+++ docs/pages/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Upgrade to tailwindcss v4 and daisyUI v5
+
 ## [3.11.0] - 2025-03-04
 
 ### Added